pyproject.toml
[tool.poetry]
name = "LogarithmPlotter"
version = "0.6.0"
description = "Create and edit Bode plots"
authors = ["Ad5001 <mail@ad5001.eu>"]
license = "GPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
PySide6-Essentials = "^6.7.2"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.10.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-qt = "^4.4.0"
stdeb = "^0.10.0"
